+++ src/Relude/Foldable/Fold.hs
@@ -177,12 +177,11 @@ Nothing
 False
 -}
 andM :: (Foldable f, Monad m) => f (m Bool) -> m Bool
-andM = go . toList
+andM = foldr go (pure True)
   where
-    go []     = pure True
-    go (p:ps) = do
+    go p acc = do
         q <- p
-        if q then go ps else pure False
+        if q then acc else pure False
 {-# INLINEABLE andM #-}
 {-# SPECIALIZE andM :: [IO Bool] -> IO Bool #-}
 
@@ -196,12 +195,11 @@ Just True
 Nothing
 -}
 orM :: (Foldable f, Monad m) => f (m Bool) -> m Bool
-orM = go . toList
+orM = foldr go (pure False)
   where
-    go []     = pure False
-    go (p:ps) = do
+    go p acc = do
         q <- p
-        if q then pure True else go ps
+        if q then pure True else acc
 {-# INLINEABLE orM #-}
 {-# SPECIALIZE orM  :: [IO Bool] -> IO Bool #-}
 
@@ -215,12 +213,11 @@ Just False
 Nothing
 -}
 allM :: (Foldable f, Monad m) => (a -> m Bool) -> f a -> m Bool
-allM p = go . toList
+allM p = foldr go (pure True)
   where
-    go []     = pure True
-    go (x:xs) = do
+    go x acc = do
         q <- p x
-        if q then go xs else pure False
+        if q then acc else pure False
 {-# INLINEABLE allM #-}
 {-# SPECIALIZE allM :: (a -> IO Bool) -> [a] -> IO Bool #-}
 
@@ -234,12 +231,11 @@ Just True
 Nothing
 -}
 anyM :: (Foldable f, Monad m) => (a -> m Bool) -> f a -> m Bool
-anyM p = go . toList
+anyM p = foldr go (pure False)
   where
-    go []     = pure False
-    go (x:xs) = do
+    go x acc = do
         q <- p x
-        if q then pure True else go xs
+        if q then pure True else acc
 {-# INLINEABLE anyM #-}
 {-# SPECIALIZE anyM :: (a -> IO Bool) -> [a] -> IO Bool #-}